# quinte-backend-service/openapi.yaml
openapi: 3.1.0
info:
  title: Quinte Backend Service API
  version: 1.0.0
  description: Student management CRUD API.
servers:
  - url: http://localhost:8080
paths:
  /api/students:
    post:
      operationId: createStudent
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: listStudents
      summary: List students with pagination, sorting, and search
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            default: id
            enum: [id, name, email, createdDate]
        - in: query
          name: sortDir
          required: false
          schema:
            type: string
            default: asc
            enum: [asc, desc]
        - in: query
          name: search
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paged students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentPageResponse'
        '400':
          description: Invalid query params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/{id}:
    get:
      operationId: getStudentById
      summary: Get student by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateStudent
      summary: Update existing student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdateRequest'
      responses:
        '200':
          description: Updated student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteStudent
      summary: Soft delete student by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Delete success message
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  message:
                    type: string
                required: [message]
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    StudentCreateRequest:
      type: object
      additionalProperties: false
      required: [name, email, phoneNumber]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        phoneNumber:
          type: string
          pattern: '^[0-9]{10,15}$'
    StudentUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        phoneNumber:
          type: string
          pattern: '^[0-9]{10,15}$'
    StudentResponse:
      type: object
      additionalProperties: false
      required: [id, name, email, phoneNumber, createdDate]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        createdDate:
          type: string
          format: date-time
    StudentPageResponse:
      type: object
      additionalProperties: true
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponse'
        pageable:
          type: object
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        sort:
          type: object
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: object
          additionalProperties:
            type: string